+++ cpp/dolfinx/refinement/PlazaRefinementND.cpp
@@ -65,7 +65,7 @@ mesh::Mesh compute_refinement(const mesh::Mesh& mesh, ParallelRefinement& p_ref,
 
   // Make new vertices in parallel
   p_ref.create_new_vertices();
-  const std::map<std::size_t, std::size_t>& new_vertex_map
+  const std::map<std::int32_t, std::int64_t>& new_vertex_map
       = p_ref.edge_to_new_vertex();
 
   std::vector<std::size_t> parent_cell;
